Use cleanup_helper fixture in integration testing
The fixture allows cleaning up aliases, keybindings, ... after testing
and was previously only used for end2end testing. It is also useful for
the integration tests though, to keep the state clean and remove cleanup
logic from the individual tests.
